requirements.txt
langchain
langchain_community
crewai
crewai_tools
streamlit
ollama
google-generativeai
python-dotenv